requirements.txt
Django
dj-database-url
dj-static
django-avatar
django-bootstrap3
django-celery
django-celery-email
django-debug-toolbar
django-extensions
django-stronghold
gunicorn
honcho
kombu
psycopg2
pytz
redis
django-redis
django-tastypie
django-braces
